+++ packages/stripe_platform_interface/lib/src/models/payment_sheet.dart
@@ -11,6 +11,7 @@ part 'payment_sheet.g.dart';
 /// For more info about the payment sheet see: https://stripe.com/docs/payments/accept-a-payment.
 @freezed
 class SetupPaymentSheetParameters with _$SetupPaymentSheetParameters {
+  @JsonSerializable(explicitToJson: true)
   const factory SetupPaymentSheetParameters({
     /// Whether or not to use a custom flow.
     ///
@@ -56,6 +57,8 @@ class SetupPaymentSheetParameters with _$SetupPaymentSheetParameters {
     bool? applePay,
 
     /// Style options for colors in PaymentSheet
+    ///
+    /// Parts can be overridden by [appearance].
     @JsonKey(toJson: UserInterfaceStyleKey.toJson) ThemeMode? style,
 
     /// Configuration related to Google Pay
@@ -68,11 +71,10 @@ class SetupPaymentSheetParameters with _$SetupPaymentSheetParameters {
     /// Defaul value is false.
     @Default(false) bool allowsDelayedPaymentMethods,
 
-    /// Button color of the checkoutButton
+    /// Appearance of the paymentsheet.
     ///
-    /// Make sure that there is enough contrast with the rest of the paymentsheet.
-    @JsonKey(toJson: ColorKey.toJson, fromJson: ColorKey.fromJson)
-        Color? primaryButtonColor,
+    /// When no appearance defined it will fallback to [style] or Stripe default.
+    PaymentSheetAppearance? appearance,
 
     /// Flag for using the test environment
     @Default(false) bool testEnv,
@@ -91,6 +93,205 @@ class SetupPaymentSheetParameters with _$SetupPaymentSheetParameters {
 
   factory SetupPaymentSheetParameters.fromJson(Map<String, dynamic> json) =>
       _$SetupPaymentSheetParametersFromJson(json);
+}
+
+/// Parameters that change the appearance of the payment sheet.
+@freezed
+class PaymentSheetAppearance with _$PaymentSheetAppearance {
+  @JsonSerializable(explicitToJson: true)
+  const factory PaymentSheetAppearance({
+    /// Color parameters
+    PaymentSheetAppearanceColors? colors,
+
+    /// Shapes parameters
+    PaymentSheetShape? shapes,
+
+    /// PaymentSheet appearance
+    PaymentSheetPrimaryButtonAppearance? primaryButton,
+  }) = _PaymentSheetAppearance;
+
+  factory PaymentSheetAppearance.fromJson(Map<String, dynamic> json) =>
+      _$PaymentSheetAppearanceFromJson(json);
+}
+
+/// Color settings for the Paymentsheet.
+@freezed
+class PaymentSheetAppearanceColors with _$PaymentSheetAppearanceColors {
+  const factory PaymentSheetAppearanceColors({
+    ///  Color of the button that represents the primary action on the payment sheet.
+    ///
+    /// Make sure there is enough contrast with [background].
+    @JsonKey(toJson: ColorKey.toJson, fromJson: ColorKey.fromJson)
+        Color? primary,
+
+    /// Background color of the payment sheet.
+    ///
+    /// Make sure there is enough contrast with [primary].
+    @JsonKey(toJson: ColorKey.toJson, fromJson: ColorKey.fromJson)
+        Color? background,
+
+    /// Background color of the payment sheet components.
+    @JsonKey(toJson: ColorKey.toJson, fromJson: ColorKey.fromJson)
+        Color? componentBackground,
+
+    ///  Border color of the payment sheet components.
+    @JsonKey(toJson: ColorKey.toJson, fromJson: ColorKey.fromJson)
+        Color? componentBorder,
+
+    ///  Divider color of the payment sheet components.
+    @JsonKey(toJson: ColorKey.toJson, fromJson: ColorKey.fromJson)
+        Color? componentDivider,
+
+    /// Color of the entered text in the payment components.
+    @JsonKey(toJson: ColorKey.toJson, fromJson: ColorKey.fromJson)
+        Color? componentText,
+
+    /// Primary text color.
+    @JsonKey(toJson: ColorKey.toJson, fromJson: ColorKey.fromJson)
+        Color? primaryText,
+
+    /// Secondary text color.
+    @JsonKey(toJson: ColorKey.toJson, fromJson: ColorKey.fromJson)
+        Color? secondaryText,
+
+    /// Place holder text color.
+    @JsonKey(toJson: ColorKey.toJson, fromJson: ColorKey.fromJson)
+        Color? placeholderText,
+
+    /// Color of the displayed icons
+    @JsonKey(toJson: ColorKey.toJson, fromJson: ColorKey.fromJson) Color? icon,
+
+    /// Color of the warning and error messages.
+    @JsonKey(toJson: ColorKey.toJson, fromJson: ColorKey.fromJson) Color? error,
+  }) = _PaymentSheetAppearanceColors;
+
+  factory PaymentSheetAppearanceColors.fromJson(Map<String, dynamic> json) =>
+      _$PaymentSheetAppearanceColorsFromJson(json);
+}
+
+/// Parameters that change the appearance of the payment sheet.
+@freezed
+class PaymentSheetShape with _$PaymentSheetShape {
+  @JsonSerializable(explicitToJson: true)
+  const factory PaymentSheetShape({
+    /// Borderradius for the paymentsheet corners
+    double? borderRadius,
+
+    /// Borderwidth for the paymentsheet components
+    double? borderWidth,
+
+    /// Appearance config of the payment sheet shadow
+    PaymentSheetShadowParams? shadow,
+  }) = _PaymentSheetShape;
+
+  factory PaymentSheetShape.fromJson(Map<String, dynamic> json) =>
+      _$PaymentSheetShapeFromJson(json);
+}
+
+/// Shadow config parameters
+@freezed
+class PaymentSheetShadowParams with _$PaymentSheetShadowParams {
+  @JsonSerializable(explicitToJson: true)
+  const factory PaymentSheetShadowParams({
+    /// Shadow color
+    @JsonKey(toJson: ColorKey.toJson, fromJson: ColorKey.fromJson) Color? color,
+
+    /// Shadow opacity
+    double? opacity,
+
+    /// Shadow offset
+    PaymentSheetShadowOffset? offset,
+  }) = _PaymentSheetShadowParams;
+
+  factory PaymentSheetShadowParams.fromJson(Map<String, dynamic> json) =>
+      _$PaymentSheetShadowParamsFromJson(json);
+}
+
+/// Shadow config parameters
+@freezed
+class PaymentSheetShadowOffset with _$PaymentSheetShadowOffset {
+  @JsonSerializable(explicitToJson: true)
+  const factory PaymentSheetShadowOffset({
+    /// X value
+    double? x,
+
+    /// Y value
+    double? y,
+  }) = _PaymentSheetShadowOffset;
+
+  factory PaymentSheetShadowOffset.fromJson(Map<String, dynamic> json) =>
+      _$PaymentSheetShadowOffsetFromJson(json);
+}
+
+@freezed
+class PaymentSheetPrimaryButtonAppearance
+    with _$PaymentSheetPrimaryButtonAppearance {
+  @JsonSerializable(explicitToJson: true)
+  const factory PaymentSheetPrimaryButtonAppearance({
+    /// color theme of the primary button
+    PaymentSheetPrimaryButtonTheme? colors,
+
+    /// Shape params of the primary button
+    PaymentSheetPrimaryButtonShape? shapes,
+
+  }) = _PaymentSheetPrimaryButtonAppearance;
+
+  factory PaymentSheetPrimaryButtonAppearance.fromJson(
+          Map<String, dynamic> json) =>
+      _$PaymentSheetPrimaryButtonAppearanceFromJson(json);
+}
+@freezed
+class PaymentSheetPrimaryButtonShape with _$PaymentSheetPrimaryButtonShape {
+    @JsonSerializable(explicitToJson: true)
+  const factory PaymentSheetPrimaryButtonShape({
+    /// Configuration of the primary button's shadow.
+    PaymentSheetShadowParams? shadow,
+
+    /// the blur radius of the button
+    double? blurRadius,
+
+    /// border width of the primary button on the payment sheet
+    double? borderWidth,
+  }) = _PaymentSheetPrimaryButtonShape;
+
+  factory PaymentSheetPrimaryButtonShape.fromJson(Map<String, dynamic> json) =>
+      _$PaymentSheetPrimaryButtonShapeFromJson(json);
+}
+
+@freezed
+class PaymentSheetPrimaryButtonTheme with _$PaymentSheetPrimaryButtonTheme {
+  @JsonSerializable(explicitToJson: true)
+  const factory PaymentSheetPrimaryButtonTheme({
+    /// Colors when displaying button in dark theme
+    PaymentSheetPrimaryButtonThemeColors? dark,
+
+    /// Colors when displaying button in light theme
+    PaymentSheetPrimaryButtonThemeColors? light,
+  }) = _PaymentSheetPrimaryButtonTheme;
+
+  factory PaymentSheetPrimaryButtonTheme.fromJson(Map<String, dynamic> json) =>
+      _$PaymentSheetPrimaryButtonThemeFromJson(json);
+}
+
+@freezed
+class PaymentSheetPrimaryButtonThemeColors
+    with _$PaymentSheetPrimaryButtonThemeColors {
+  const factory PaymentSheetPrimaryButtonThemeColors({
+    /// Primary button background color
+    @JsonKey(toJson: ColorKey.toJson, fromJson: ColorKey.fromJson)
+        Color? background,
+
+    /// Primary button text color
+    @JsonKey(toJson: ColorKey.toJson, fromJson: ColorKey.fromJson) Color? text,
+
+    /// Primary button border color
+    @JsonKey(toJson: ColorKey.toJson, fromJson: ColorKey.fromJson)
+        Color? border,
+  }) = _PaymentSheetPrimaryButtonThemeColors;
+
+  factory PaymentSheetPrimaryButtonThemeColors.fromJson(
+          Map<String, dynamic> json) =>
+      _$PaymentSheetPrimaryButtonThemeColorsFromJson(json);
 }
 
 class UserInterfaceStyleKey {