+++ packages/stripe_web/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 1.1.0
+- Sdk updates
+
 ## 1.0.0
 ** Breaking changes **
 - Support Flutter 3.0.0